Move Jest mocking of IO and Monaco into Mock files
